Add javascript Class Generator and add counter in input api name
