+++ assets/js/generator-form.js
@@ -1,3 +1,1 @@
-const $ = require('jquery');
-
 require('./generator/form');
